+++ src/locale/ar/transport/index.js
@@ -1,0 +1,7 @@
+const transport = {
+    vehicleManufacturer: require('./vehicleManufacturer'),
+    vehicleModel: require('./vehicleModel'),
+    vehicleType: require('./vehicleType'),
+    vehicleFuel: require('./vehicleFuel')
+};
+module.exports = transport;