+++ src/locale/ar/vehicle/index.js
@@ -1,7 +1,0 @@
-const vehicle = {
-    fuel: require('./fuel'),
-    manufacturer: require('./manufacturer'),
-    module: require('./model'),
-    type: require('./vehicle_type'),
-};
-module.exports = vehicle;