+++ src/locale/en/airport/index.js
@@ -1,2 +1,0 @@
-const airport = { name: require('./name') };
-module.exports = airport;