+++ src/locale/en/transport/index.js
@@ -1,0 +1,8 @@
+const transport = {
+    vehicleManufacturer: require('./vehicleManufacturer'),
+    vehicleModel: require('./vehicleModel'),
+    vehicleType: require('./vehicleType'),
+    vehicleFuel: require('./vehicleFuel'),
+    airportName: require('./airportName'),
+};
+module.exports = transport;